refactoring bundler logic
